requirements.txt
watchdog>=2.2.0
requests>=2.28.0
pywin32>=300
Pillow
Pillow>=9.3.0
